Remove Flysystem v2
