--- .github/workflows/flysystemV1.yml
@@ -1,89 +1,0 @@
-name: FlysystemV1
-
-on:
-  pull_request: ~
-  push:
-    branches:
-      - master
-
-jobs:
-  phpstan:
-    name: Static checks
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Set up PHP
-        uses: shivammathur/setup-php@2.7.0
-        with:
-          php-version: 7.4
-          coverage: none
-
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Download dependencies
-        run: |
-          composer remove --no-update --no-interaction --dev league/flysystem-adapter-test-utilities
-          composer req --no-update --dev league/flysystem:^1.0
-          composer update --no-interaction --prefer-dist --optimize-autoloader
-
-      - name: Modify config
-        run: |
-          sed -i -re 's/S3FilesystemV1/S3FilesystemV2/' phpstan.neon.dist
-          sed -i -re 's/S3FilesystemV1/S3FilesystemV2/' psalm.xml
-          sed -i -re '/Integration\/Laravel\/Filesystem/d' phpstan.neon.dist
-          sed -i -re '/Integration\/Laravel\/Filesystem/d' psalm.xml
-
-      - name: PHPStan
-        uses: docker://oskarstark/phpstan-ga:0.12.48
-        with:
-          entrypoint: /composer/vendor/bin/phpstan
-          args: analyze --no-progress
-
-      - name: Psalm
-        uses: docker://vimeo/psalm-github-actions:3.17.2
-        with:
-          args: --no-progress --show-info=false --stats
-
-  phpunit:
-    name: PHPUnit
-    runs-on: ubuntu-latest
-    steps:
-      - name: Set up PHP
-        uses: shivammathur/setup-php@2.7.0
-        with:
-          php-version: 7.4
-          coverage: none
-
-      - name: Checkout code
-        uses: actions/checkout@v2
-
-      - name: Initialize tests
-        run: |
-          cd src/Integration/Flysystem/S3
-          make initialize
-
-      - name: Download phpunit
-        run: |
-          rm composer.*
-          composer init --no-interaction
-
-          echo ::group::Download
-          composer req symfony/phpunit-bridge
-          echo ::endgroup::
-
-          echo ::group::Install
-          ./vendor/bin/simple-phpunit install
-          echo ::endgroup::
-
-      - name: Download dependencies
-        run: |
-          cd src/Integration/Flysystem/S3
-          composer remove --no-update --no-interaction --dev league/flysystem-adapter-test-utilities
-          composer req --no-update --dev league/flysystem:^1.0
-          sed -i -re 's/"require": \{/"minimum-stability": "dev","prefer-stable": true,"require": \{/' composer.json
-          composer update --no-interaction --prefer-dist --optimize-autoloader
-
-      - name: Run tests
-        run: |
-          ./vendor/bin/simple-phpunit -c src/Integration/Flysystem/S3/phpunit.xml.dist
