+++ composer.json
@@ -23,8 +23,7 @@
         "illuminate/mail": "^6.18.13 || ^7.10",
         "illuminate/queue": "^6.18.13 || ^7.10",
         "illuminate/support": "^6.18.13 || ^7.10",
-        "league/flysystem": "^1.0.65 || ^2.0.0-rc.1",
-        "league/flysystem-adapter-test-utilities": "^2.0.0-rc.1",
+        "league/flysystem": "^1.0.65",
         "league/mime-type-detection": "^1.0",
         "matthiasnoback/symfony-config-test": "^4.1",
         "monolog/monolog": "^1.1 || ^2.0",
