+++ psalm.xml
@@ -8,7 +8,6 @@
     errorBaseline="psalm.baseline.xml"
 >
     <projectFiles>
-        <file name="src/Integration/Flysystem/S3/tests/version-bridge.php" />
         <directory name="src" />
         <ignoreFiles>
             <directory name="src/CodeGenerator" />
@@ -16,7 +15,6 @@
             <directory name="src/Service/*/tests" />
             <directory name="src/Integration/*/*/tests" />
             <file name="src/Core/src/HttpClient/AwsRetryStrategy.php" /><!-- # requires symfony/http-client: 5.2-->
-            <file name="src/Integration/Flysystem/S3/src/S3FilesystemV1.php" />
             <file name="src/Integration/Laravel/Filesystem/src/AsyncAwsFilesystemManager.php" />
             <file name="src/Integration/Laravel/Filesystem/src/AsyncAwsFilesystemAdapter.php" />
             <directory name="vendor" />
