+++ src/Integration/Flysystem/S3/README.md
@@ -3,7 +3,8 @@
 ![](https://github.com/async-aws/flysystem-s3/workflows/Tests/badge.svg?branch=master)
 ![](https://github.com/async-aws/flysystem-s3/workflows/BC%20Check/badge.svg?branch=master)
 
-Flysystem S3 adapter for Flysystem version 1 and 2.
+Flysystem S3 adapter for Flysystem version 1. The adapter for Flysystem v2 is found
+here: https://flysystem.thephpleague.com/v2/docs/adapter/async-aws-s3/
 
 ## Install
