+++ src/Integration/Flysystem/S3/composer.json
@@ -20,17 +20,16 @@
     ],
     "require": {
         "php": "^7.2.5 || ^8.0",
-        "async-aws/core": "^1.0",
-        "async-aws/s3": "^1.0",
-        "league/flysystem": "^1.0.65 || ^2.0.0-rc.1",
+        "async-aws/s3": "^1.5",
+        "league/flysystem": "^1.0.65",
         "league/mime-type-detection": "^1.0"
     },
     "require-dev": {
-        "league/flysystem-adapter-test-utilities": "^2.0.0-rc.1"
+        "symfony/phpunit-bridge": "^5.1"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.5-dev"
+            "dev-master": "1.0-dev"
         }
     },
     "autoload": {
