+++ src/Integration/Flysystem/S3/phpunit.xml.dist
@@ -3,7 +3,7 @@
          xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.3/phpunit.xsd"
          backupGlobals="false"
          colors="true"
-         bootstrap="tests/bootstrap.php"
+         bootstrap="vendor/autoload.php"
          failOnRisky="true"
          failOnWarning="true"
 >
