--- src/Integration/Flysystem/S3/tests/bootstrap.php
@@ -1,8 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-require_once \dirname(__DIR__) . '/vendor/autoload.php';
-
-require_once __DIR__ . '/version-bridge.php';
-require_once __DIR__ . '/version-bridge-tests.php';
